EES-4376 Split MethodologyStatus auth handlers into separate files
