+++ src/GovUk.Education.ExploreEducationStatistics.Admin/Security/AuthorizationHandlers/MarkMethodologyAsHigherLevelReviewAuthorizationHandler.cs
@@ -1,0 +1,73 @@
+#nullable enable
+using System.Threading.Tasks;
+using GovUk.Education.ExploreEducationStatistics.Common.Services.Security;
+using GovUk.Education.ExploreEducationStatistics.Content.Model;
+using GovUk.Education.ExploreEducationStatistics.Content.Model.Repository.Interfaces;
+using Microsoft.AspNetCore.Authorization;
+using static GovUk.Education.ExploreEducationStatistics.Admin.Security.SecurityClaimTypes;
+using static GovUk.Education.ExploreEducationStatistics.Admin.Security.AuthorizationHandlers.AuthorizationHandlerResourceRoleService;
+using static GovUk.Education.ExploreEducationStatistics.Common.Services.CollectionUtils;
+using static GovUk.Education.ExploreEducationStatistics.Content.Model.MethodologyApprovalStatus;
+using static GovUk.Education.ExploreEducationStatistics.Content.Model.PublicationRole;
+
+namespace GovUk.Education.ExploreEducationStatistics.Admin.Security.AuthorizationHandlers;
+// ReSharper disable once ClassNeverInstantiated.Global
+public class MarkMethodologyAsHigherLevelReviewRequirement : IAuthorizationRequirement
+{
+}
+
+public class MarkMethodologyAsHigherLevelReviewAuthorizationHandler : AuthorizationHandler<
+    MarkMethodologyAsHigherLevelReviewRequirement, MethodologyVersion>
+{
+    private readonly IMethodologyVersionRepository _methodologyVersionRepository;
+    private readonly IMethodologyRepository _methodologyRepository;
+    private readonly AuthorizationHandlerResourceRoleService _authorizationHandlerResourceRoleService;
+
+    public MarkMethodologyAsHigherLevelReviewAuthorizationHandler(
+        IMethodologyVersionRepository methodologyVersionRepository,
+        IMethodologyRepository methodologyRepository,
+        AuthorizationHandlerResourceRoleService authorizationHandlerResourceRoleService)
+    {
+        _methodologyVersionRepository = methodologyVersionRepository;
+        _methodologyRepository = methodologyRepository;
+        _authorizationHandlerResourceRoleService = authorizationHandlerResourceRoleService;
+    }
+
+    protected override async Task HandleRequirementAsync(AuthorizationHandlerContext context,
+        MarkMethodologyAsHigherLevelReviewRequirement requirement,
+        MethodologyVersion methodologyVersion)
+    {
+        // If the Methodology is already public, it cannot be marked for higher level review
+        if (await _methodologyVersionRepository.IsPubliclyAccessible(methodologyVersion.Id))
+        {
+            return;
+        }
+
+        if (SecurityUtils.HasClaim(context.User, SubmitAllMethodologiesToHigherReview))
+        {
+            context.Succeed(requirement);
+            return;
+        }
+
+        var allowedPublicationRoles = methodologyVersion.Status == Approved
+            ? ListOf(Approver)
+            : ListOf(Owner, Approver);
+
+        var allowedReleaseRoles = methodologyVersion.Status == Approved
+            ? ListOf(ReleaseRole.Approver)
+            : ReleaseEditorAndApproverRoles;
+
+        var owningPublication =
+            await _methodologyRepository.GetOwningPublication(methodologyVersion.MethodologyId);
+
+        if (await _authorizationHandlerResourceRoleService
+                .HasRolesOnPublicationOrLatestRelease(
+                    context.User.GetUserId(),
+                    owningPublication.Id,
+                    allowedPublicationRoles,
+                    allowedReleaseRoles))
+        {
+            context.Succeed(requirement);
+        }
+    }
+}
